Homework 6 -Packages added -toString only once. I need to rebuild project to apply 5 toString methods.
